Cyclic dependencies progress
