+++ CSC8508CoreClasses/IComponent.h
@@ -8,7 +8,6 @@
 #include "Transform.h"
 #include "GameObject.h"
 #include "ISerializable.h"
-
 #include <unordered_set>
 
 namespace NCL::CSC8508 
@@ -20,7 +19,6 @@ namespace NCL::CSC8508
 		IComponent(GameObject& gameObject) : gameObject(gameObject), enabled(true) {}
 
 		virtual ~IComponent() = default;
-
 
 		/**
 		 * Function invoked after the object and components have been instantiated.
@@ -85,36 +83,9 @@ namespace NCL::CSC8508
 			return GetDerivedTypes().count(std::type_index(typeInfo)) > 0;
 		}
 
-
-		struct ComponentDataStruct : public ISerializable {
-			ComponentDataStruct() : enabled(1) {}
-			ComponentDataStruct(bool enabled) : enabled(enabled) {}
-
-			bool enabled;
-
-			static auto GetSerializedFields() {
-				return std::make_tuple(
-					SERIALIZED_FIELD(ComponentDataStruct, enabled)
-				);
-			}
-		};
-
-		void Load(std::string folderPath, std::string name) override {
-
-			ComponentDataStruct loadedSaveData = ISerializedData::LoadISerializable<ComponentDataStruct>(folderPath, name);
-			std::cout << loadedSaveData.enabled << ": Component is enabled" << std::endl;
-		}
-
-		std::string Save(std::string folderPath) override
-		{
-			int id = 6;
-			std::string fileName = "game_data%" + std::to_string(id) + ".gdmt";
-			ComponentDataStruct saveInfo(enabled);
-			SaveManager::GameData saveData = ISerializedData::CreateGameData<ComponentDataStruct>(saveInfo);
-			SaveManager::SaveGameData(fileName, saveData);
-			return fileName;
-		}
-
+		void Load(std::string folderPath, std::string name) override; 
+		std::string Save(std::string folderPath) override;
+		struct ComponentDataStruct;
 
 	protected:
 		virtual void OnAwake() {}
