+++ CSC8508CoreClasses/INetworkDeltaComponent.h
@@ -5,11 +5,6 @@
 #ifndef INETWORKDELTACOMPONENT_H
 #define INETWORKDELTACOMPONENT_H
 
-#include "Transform.h"
-#include "IComponent.h"
-#include "NetworkBase.h"
-#include "Event.h"
-#include "EventManager.h"
 #include "INetworkComponent.h"
 using std::vector;
